test/pyproject.toml
[project]
name = "rag-api"
version = "1.0.0"
description = "RAG API with pgvector"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "psycopg2-binary>=2.9.0",
    "langchain-text-splitters>=0.2.0",
    "langchain-huggingface>=0.0.1",
    "langchain-core>=0.2.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "pypdf>=3.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",       # for TestClient
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
